Added new clock time management system: RTC stores UTC and local time is determined from a configurable timezone setting (including adjustments for daylight saving time).  Unix timestamp is now core time format.  Coordinator node will now regularly poll an NTP server and will frequently broadcast time over XBee network for other nodes to sync their clocks.  Other code cleanup and bug fixes.

+++ Software/PODD_Firmware_v2/SensorPod_FW/pod_serial.cpp
@@ -404,7 +404,7 @@ int serialIntegerPrompt(String prompt, bool reprompt, int default0) {
    is received.
    A default value can be specified, used if an empty
    response is given. */
-int serialLongPrompt(String prompt, bool reprompt, long default0) {
+long serialLongPrompt(String prompt, bool reprompt, long default0) {
   while (true) {
     // Provide prompt to user
     Serial.print(prompt);
